feat: Update users-index page filter by role
Updated the users-index page to add a filter option by role. Implemented the filterUsers function to filter users based on the selected role. Added a ChoiceSelect component to allow the user to select a role and updated the TableList component to display filtered users. Also made modifications to the utils.ts file to include translations for role options.