Refactor UserController's destroy method to remove multiple users at once
The UserController's destroy method was refactored to allow the removal of multiple users at once. The method now expects a list of user IDs to be passed in the request, validates the input, and deletes the corresponding users from the database. The route was also updated to reflect this change.